add an option to pre-open or close the popup
To make popup blockers happy, we need to open the window in a synchronous
matter, inside a click handler.  This is usually fine, but if we need to
generate some data asynchronously (like the challenges needed for
`requestLogin`), popup will get blocked.  To avoid this, we can `open()` it in
the click handler and then continue the communication whenever we want.