test: Add unofficial support for `podman`
Podman is an open-source daemonless way to run containers in Linux.

Essentially, this means it doesn't need `root` access to run,
so it's slightly more secure.

Its CLI is so similar to `docker`, that you can just use
`alias docker=podman` and it should work in 99.99% of cases.

However, `alias docker=podman` doesn't work in bash scripts
by default, so I've added a flag `-D PATH` to set the docker path,
which also allows replacing it with `podman`.